+++ src/shared/validation/validators/age.validator.ts
@@ -1,5 +1,5 @@
 import { ValidateFunction, validate } from '@shared/validation/lib/validate'
-import { positiveNumberSchema } from '../validation/schemas/positive-number.schema'
+import { positiveNumberSchema } from '../schemas/positive-number.schema'
 
 export const validateAge: ValidateFunction<number> = (value) =>
   validate(positiveNumberSchema, value, 'positive number')
