+++ src/shared/validation/validators/email.validator.ts
@@ -1,5 +1,5 @@
 import { ValidateFunction, validate } from '@shared/validation/lib/validate'
-import { eMailSchema } from '../validation/schemas/email.schema'
+import { eMailSchema } from '../schemas/email.schema'
 
 export const validateEMail: ValidateFunction<string> = (value) =>
   validate(eMailSchema, value, 'valid email')
