+++ src/shared/validation/validators/picture.validator.ts
