+++ src/shared/validation/validators/terms.validator.ts
